update flow layout section controller
